move from class to function
